LayoutTests/imported/w3c/web-platform-tests/clipboard-apis/detached-iframe/clipboard-on-detached-iframe.https-expected.txt


FAIL Verify navigator.clipboard attached only on attached frames assert_equals: detached child iframe's clipboard should not exist expected null but got object "[object Clipboard]"
